Add files via upload
Data Visualization of Analysing the IPL data
The project is the analysis of the IPL data over all seasons played. In this project an attemp is made to
make an interactive vusualisation that helps in making the data more sense even for a comman audience.
Here I have made many plots of various categories that are analysed from the data. The data is collected
from the kaggle website. The kaggle website is a platform where many kinds of data are available for free.
Anyone interested in looking at data can download it. It also has data in many formats such as CSV, YAML,
JSON, etc. You can access the data here: https://www.kaggle.com/manasgarg/ipl
The format of the data is CSV. The data is of two files and named as “deliveries.csv” and “matches.csv”.
The deliveries.csv has data about each ball that is being bowled in the IPL so far, and it has data with 21
variables. The matches.csv has the data about the match which constitutes 18 variables. The IPL is a series
that are played between 13 teams in India, and it has been played for 12 seasons by now. The data has
information about all the players who are part of it and the teams that won in each match. It also contains
the data like the number of sixes, fours, wides, wickets information of each ball categorized by the match.
The data contains information about where the match is being held and the date.
A shiny server is also made using the data, where we can change the variables and the plot is shown for each
different variable. Shiny App has three parts, named as UI(User Interface), server and Shiny App. The UI is
used to design the way the app is being visible to all. The server is used to plot the information we want.
The Shiny App visualizes the final output of the UI and server together.